Moved test to correct folder
