+++ client/my-sites/subscribers/components/subscriber-list/test/subscriber-row.test.tsx
@@ -6,19 +6,26 @@ import { isEnabled } from '@automattic/calypso-config';
 import { render, fireEvent, screen } from '@testing-library/react';
 import React from 'react';
 import useSubscriptionPlans from '../../../hooks/use-subscription-plans';
-import { Subscriber } from '../../../types';
 import { SubscriberRow } from '../subscriber-row';
 
 jest.mock( '@automattic/calypso-config' );
 jest.mock( '../../../hooks/use-subscription-plans' );
 
 describe( 'SubscriberRow', () => {
-	const mockSubscriber: Partial< Subscriber > = {
-		avatar: 'avatar.jpg',
-		display_name: 'John Doe',
-		email_address: 'john@example.com',
-		date_subscribed: new Date().toISOString(),
-		open_rate: 75,
+	const commonProps: any = {
+		user_id: 123,
+		subscription_id: 456,
+		subscriptions: [],
+		locale: 'en',
+		onUnsubscribe: jest.fn(),
+		onView: jest.fn(),
+		subscriber: {
+			date_subscribed: new Date( '2023-01-01' ).toISOString(),
+			avatar: 'http://example.com/avatar.png',
+			display_name: 'Test User',
+			email_address: 'test@example.com',
+			open_rate: 50,
+		},
 	};
 
 	const mockOnUnsubscribe = jest.fn();
@@ -32,11 +39,7 @@ describe( 'SubscriberRow', () => {
 		( isEnabled as jest.MockedFunction< typeof isEnabled > ).mockReturnValue( true );
 
 		render(
-			<SubscriberRow
-				subscriber={ mockSubscriber as Subscriber }
-				onView={ mockOnView }
-				onUnsubscribe={ mockOnUnsubscribe }
-			/>
+			<SubscriberRow { ...commonProps } onView={ mockOnView } onUnsubscribe={ mockOnUnsubscribe } />
 		);
 	} );
 
@@ -45,8 +48,8 @@ describe( 'SubscriberRow', () => {
 	} );
 
 	it( 'should display the correct subscriber data', () => {
-		expect( screen.getByText( 'John Doe' ) ).toBeInTheDocument();
-		expect( screen.getByText( 'john@example.com' ) ).toBeInTheDocument();
+		expect( screen.getByText( commonProps.subscriber.display_name ) ).toBeInTheDocument();
+		expect( screen.getByText( commonProps.subscriber.email_address ) ).toBeInTheDocument();
 		expect( screen.getByText( 'Plan 1' ) ).toBeInTheDocument();
 		expect( screen.getByText( 'Plan 2' ) ).toBeInTheDocument();
 	} );
@@ -58,7 +61,7 @@ describe( 'SubscriberRow', () => {
 		const viewButton = await screen.findByRole( 'menuitem', { name: /view/i } );
 		fireEvent.click( viewButton );
 
-		expect( mockOnView ).toHaveBeenCalledWith( mockSubscriber );
+		expect( mockOnView ).toHaveBeenCalledWith( commonProps.subscriber );
 	} );
 
 	it( 'should trigger onUnsubscribe when unsubscribe button is clicked', async () => {
@@ -68,6 +71,19 @@ describe( 'SubscriberRow', () => {
 		const unsubscribeButton = await screen.findByRole( 'menuitem', { name: /unsubscribe/i } );
 		fireEvent.click( unsubscribeButton );
 
-		expect( mockOnUnsubscribe ).toHaveBeenCalledWith( mockSubscriber );
+		expect( mockOnUnsubscribe ).toHaveBeenCalledWith( commonProps.subscriber );
+	} );
+
+	it( 'should render the subscriber profile correctly', () => {
+		expect( screen.getByText( commonProps.subscriber.display_name ) ).toBeInTheDocument();
+		expect( screen.getByText( commonProps.subscriber.email_address ) ).toBeInTheDocument();
+		expect( screen.getByAltText( 'Profile pic' ) ).toHaveAttribute(
+			'src',
+			commonProps.subscriber.avatar
+		);
+	} );
+
+	it( 'should display the date in the correct format', () => {
+		expect( screen.getByText( 'January 1, 2023' ) ).toBeInTheDocument();
 	} );
 } );
