+++ client/my-sites/subscribers/components/subscriber-profile/test/subscriber-profile.test.tsx
@@ -3,7 +3,7 @@
  */
 import { render, screen } from '@testing-library/react';
 import React from 'react';
-import { SubscriberProfile } from '../../components/subscriber-profile';
+import SubscriberProfile from '../subscriber-profile';
 
 const testName = 'Test Name';
 const testEmail = 'test@test.com';
